Created user authentication

+++ server/src/routes/memories.ts
@@ -3,8 +3,13 @@ import { z } from "zod";
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify();
+  });
+
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: { userId: request.user.sub },
       orderBy: { createdAt: "asc" },
     });
 
@@ -15,7 +20,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     }));
   });
 
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, response) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -24,6 +29,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
     const memory = await prisma.memory.findUniqueOrThrow({
       where: { id },
     });
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return response.status(401).send();
+    }
 
     return memory;
   });
@@ -41,14 +50,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "eae892da-dc76-41f4-8685-f14cc2fdf89f",
+        userId: request.user.sub,
       },
     });
 
     return memory;
   });
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, response) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -60,7 +69,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
     const { id } = paramsSchema.parse(request.params);
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: { id },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return response.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
       where: { id },
       data: {
         content,
@@ -72,11 +89,19 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory;
   });
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, response) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
     const { id } = paramsSchema.parse(request.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: { id },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return response.status(401).send();
+    }
 
     await prisma.memory.delete({
       where: { id },
